design changes
I fixed a lot of the formatting and design issues such as the song cover and text moving and altering whenever a new song appeared. I made these components static to allow for a more user friendly interface. I also removed the border around the picture, changed the background color to dark grey to make it easier on the eyes, fixed the select menu problem that kept overlapping with the picture, and moved it to a better spot (up for debate). I then implemented button icons to make it look nicer.